Added designation column in user model and added treeView dropdown in employee add view

+++ database/migrations/2021_09_10_044336_create_offices_table.php
@@ -16,7 +16,7 @@ class CreateOfficesTable extends Migration
         Schema::create('offices', function (Blueprint $table) {
             $table->id();
             $table->foreignId('business_id')->constrained();
-            $table->unsignedBigInteger('parent_id')->default(0);
+            $table->integer('parent_id')->default(0);
             $table->string('name');
             $table->string('email');
             $table->string('address');
